+++ app/src/main/java/com/example/week12/Presentation/Models/BooksNetworkItem.kt
@@ -4,6 +4,6 @@ data class BooksNetworkItem(
     val title: String = "",
     val description: String = "",
     val image: String = "",
-    val rating: Int = 0,
+    val rating: Double = 0.0,
     val publishedAt: String = ""
 )
